# specs/001-project-management-oa/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 项目管理OA系统 API
  description: 基于Excel需求的项目管理OA系统API文档
  version: 1.0.0
  contact:
    name: 开发团队
    email: dev@company.com
servers:
  - url: http://localhost:8080/api/v1
    description: 开发环境
  - url: https://api.project-oa.com/api/v1
    description: 生产环境
security:
  - BearerAuth: []
paths:
...
  /projects/{id}/contracts:
    get:
      tags:
        - 合同管理
      summary: 获取项目合同列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 合同列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contract'
        '404':
          description: 项目不存在

    post:
      tags:
        - 合同管理
      summary: 创建合同
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '404':
          description: 项目不存在

  # 合同补充协议管理
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        contract_number:
          type: string
        contract_type:
          type: string
        sign_date:
          type: string
          format: date
        contract_rate:
          type: number
          format: float
        contract_amount:
          type: number
          format: float
          description: 合同金额（应等于设计费+勘察费+咨询费之和）
        design_fee:
          type: number
          format: float
          description: 设计费
        survey_fee:
          type: number
          format: float
          description: 勘察费
        consultation_fee:
          type: number
          format: float
          description: 咨询费
        file_path:
          type: string
        project_id:
          type: integer
        amendments:
          type: array
          items:
            $ref: '#/components/schemas/ContractAmendment'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ContractAmendment:
      type: object
      properties:
        id:
          type: integer
        amendment_number:
          type: string
        sign_date:
          type: string
          format: date
        file_path:
          type: string
        description:
          type: string
        contract_id:
          type: integer
        contract:
          $ref: '#/components/schemas/Contract'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateContractRequest:
      type: object
      required:
        - contract_number
        - contract_type
        - sign_date
        - contract_amount
      properties:
        contract_number:
          type: string
        contract_type:
          type: string
        sign_date:
          type: string
          format: date
        contract_rate:
          type: number
          format: float
        contract_amount:
          type: number
          format: float
        design_fee:
          type: number
          format: float
          description: 设计费
        survey_fee:
          type: number
          format: float
          description: 勘察费
        consultation_fee:
          type: number
          format: float
          description: 咨询费
        file_path:
          type: string
...